requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aioredis==2.0.1
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
async-timeout==4.0.3
attrs==25.3.0
bcrypt==4.3.0
certifi==2025.4.26
charset-normalizer==3.4.1
click==8.1.8
dataclasses-json==0.6.7
distlib==0.3.9
distro==1.9.0
dnspython==2.7.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.12
fastapi-utils==0.8.0
filelock==3.18.0
frozenlist==1.6.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
jiter==0.9.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.24
langchain-community==0.3.23
langchain-core==0.3.56
langchain-text-splitters==0.3.8
langsmith==0.3.39
marshmallow==3.26.1
motor==3.7.0
multidict==6.4.3
mypy_extensions==1.1.0
numpy==2.0.2
openai==1.76.2
orjson==3.10.18
packaging==24.2
passlib==1.7.4
pipenv==2022.4.8
platformdirs==4.3.7
propcache==0.3.1
psutil==5.9.8
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
pyhumps==3.8.0
PyJWT==2.10.1
pymongo==4.12.1
pypdf==5.4.0
PyPDF2==3.0.1
python-dotenv==1.1.0
python-multipart==0.0.20
PyYAML==6.0.2
redis==6.0.0
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.2
tenacity==9.1.2
tiktoken==0.9.0
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
virtualenv==20.30.0
virtualenv-clone==0.5.7
yarl==1.20.0
zstandard==0.23.0
